refactor: 프로젝트 패키지 구조 재구성 및 도메인 저장소 인터페이스 분리
