+++ app/src/main/java/woowacourse/shopping/domain/repository/ProductRepository.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.domain.repository
+
+import woowacourse.shopping.domain.Product
+import java.util.UUID
+
+interface ProductRepository {
+    suspend fun getProducts(page: Int, pageSize: Int): List<Product>
+    suspend fun getProductById(id: UUID): Product?
+}
