+++ app/src/main/java/woowacourse/shopping/domain/repository/RecentProductRepository.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.domain.repository
+
+import kotlinx.coroutines.flow.Flow
+import woowacourse.shopping.domain.Product
+
+interface RecentProductRepository {
+    val recentProducts: Flow<List<Product>>
+    suspend fun addRecentProduct(product: Product)
+}
