--- app/src/main/java/woowacourse/shopping/repository/productRepository/RecentProductRepository.kt
@@ -1,9 +1,0 @@
-package woowacourse.shopping.repository.productRepository
-
-import kotlinx.coroutines.flow.Flow
-import woowacourse.shopping.domain.Product
-
-interface RecentProductRepository {
-    val recentProducts: Flow<List<Product>>
-    suspend fun addRecentProduct(product: Product)
-}
