+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.ui.cart
+
+import androidx.lifecycle.ViewModel
+
+class CartViewModel: ViewModel() {
+
+    companion object {
+
+    }
+}
