+++ app/src/main/java/woowacourse/shopping/ui/common/item/AddCircleButton.kt
@@ -1,0 +1,38 @@
+package woowacourse.shopping.ui.common.item
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import woowacourse.shopping.R
+
+@Composable
+fun AddCircleButton(
+    onAdd: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Surface(
+        modifier = modifier
+            .size(40.dp)
+            .clickable { onAdd() },
+        shape = CircleShape,
+        color = Color.White,
+        shadowElevation = 4.dp
+    ) {
+        Box(contentAlignment = Alignment.Center) {
+            Icon(
+                painter = painterResource(R.drawable.ic_add),
+                contentDescription = null,
+                tint = Color.Black
+            )
+        }
+    }
+}
