+++ app/src/main/java/woowacourse/shopping/ui/common/item/ProductImage.kt
@@ -1,0 +1,17 @@
+package woowacourse.shopping.ui.common.item
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import coil3.compose.AsyncImage
+
+@Composable
+fun ProductImage(
+    imageUri: String,
+    modifier: Modifier = Modifier,
+) {
+    AsyncImage(
+        model = imageUri,
+        contentDescription = "상품 사진",
+        modifier = modifier,
+    )
+}
