--- app/src/main/java/woowacourse/shopping/ui/component/item/AmountModifyButton.kt
@@ -1,86 +1,0 @@
-package woowacourse.shopping.ui.component.item
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-
-@Composable
-fun AmountModifyButton(
-    onIncrease: () -> Unit,
-    onDecrease: () -> Unit,
-    modifier: Modifier = Modifier,
-    amount: Int = 0,
-) {
-    Row(
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically,
-        modifier = modifier
-            .size(width = 126.dp, height = 42.dp)
-            .background(color = Color.White)
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxHeight()
-                .weight(1f)
-                .clickable(onClick = onDecrease),
-            contentAlignment = Alignment.Center
-        ) {
-            Text(
-                text = "-",
-                fontSize = 20.sp,
-                color = Color.Black
-            )
-        }
-
-        Box(
-            modifier = Modifier
-                .fillMaxHeight()
-                .weight(1f),
-            contentAlignment = Alignment.Center
-        ) {
-            Text(
-                text = amount.toString(),
-                fontSize = 16.sp,
-                fontWeight = FontWeight.Medium,
-                color = Color.Black
-            )
-        }
-
-        Box(
-            modifier = Modifier
-                .fillMaxHeight()
-                .weight(1f)
-                .clickable(onClick = onIncrease),
-            contentAlignment = Alignment.Center
-        ) {
-            Text(
-                text = "+",
-                fontSize = 20.sp,
-                color = Color.Black
-            )
-        }
-    }
-}
-
-@Preview
-@Composable
-private fun AmountModifyButtonPreview() {
-    AmountModifyButton(
-        onIncrease = {},
-        onDecrease = {},
-        amount = 1
-    )
-}
