--- app/src/main/java/woowacourse/shopping/ui/component/item/ProductImage.kt
@@ -1,17 +1,0 @@
-package woowacourse.shopping.ui.component.item
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import coil3.compose.AsyncImage
-
-@Composable
-fun ProductImage(
-    imageUri: String,
-    modifier: Modifier = Modifier,
-) {
-    AsyncImage(
-        model = imageUri,
-        contentDescription = "상품 사진",
-        modifier = modifier,
-    )
-}
